fix: bugs

+++ modules/spec_core/fnc/fn_core_handleVehicleRespawn.sqf
@@ -2,13 +2,11 @@ params [
     ["_vehicle", objNull, [objNull]]
 ];
 
-switch (typeOf _vehicle) do {
-    case KPLIB_preset_addVehF: {
-        [{
-            deleteVehicle (_this select 0);
-            [] call SPEC_fnc_core_spawnStartVeh;
-        }, [_vehicle], SPEC_param_startVehicleRespawn] call CBA_fnc_waitAndExecute;
-    };
+if (!isNil "KPLIB_preset_addVehF" && typeOf _vehicle isEqualTo KPLIB_preset_addVehF) then {
+    [{
+        deleteVehicle (_this select 0);
+        [] call SPEC_fnc_core_spawnStartVeh;
+    }, [_vehicle], SPEC_param_startVehicleRespawn] call CBA_fnc_waitAndExecute;
 };
 
 true
